fix: Configure ESAPI extension in final docker image build
The `LUCEE_EXTENSIONS` env var needs to be configured in the final build docker image. Currently, the way the Dockerfile is structured into two separate images causes the installed extension to be "dropped" / left behind as the docker image is rebuilt on the new context.

It may be possible instead to use a `COPY -form=build` line to copy the extension `.lex` file to the new image just like the commandbox modules. So far, I couldn't get that working.